+++ internal/resources/resource_consent_group_test.go
@@ -60,7 +60,7 @@ func TestAccConsentGroupResource_Basic(t *testing.T) {
 func testAccConsentGroupResourceConfig(groupName, description string) string {
 	return fmt.Sprintf(`
 	provider "cidaas" {
-		base_url = "https://kube-nightlybuild-dev.cidaas.de"
+		base_url = "https://automation-test.dev.cidaas.eu"
 	}
 	resource "cidaas_consent_group" "example" {
 		group_name  = "%s"
@@ -136,7 +136,7 @@ func TestAccConsentGroupResource_MissingRequired(t *testing.T) {
 			{
 				Config: `
 				provider "cidaas" {
-					base_url = "https://kube-nightlybuild-dev.cidaas.de"
+					base_url = "https://automation-test.dev.cidaas.eu"
 				}
 				resource "cidaas_consent_group" "example" {
 					description = "test description"