+++ internal/resources/resource_social_provider_test.go
@@ -2,86 +2,82 @@ package resources_test
 
 import (
 	"fmt"
-	"os"
 	"regexp"
-	"strconv"
 	"testing"
 
-	"github.com/Cidaas/terraform-provider-cidaas/helpers/cidaas"
 	acctest "github.com/Cidaas/terraform-provider-cidaas/internal/test"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
 
-const (
-	resourceSocialPorvider = "cidaas_social_provider.example"
-)
+// const (
+// 	resourceSocialPorvider = "cidaas_social_provider.example"
+// )
 
 var (
-	spName         = acctest.RandString(10)
-	spProviderName = "google"
+	spName = acctest.RandString(10)
+	// spProviderName = "google"
 	spClientID     = acctest.RandString(10)
 	spClientSecret = acctest.RandString(10)
 )
 
 // create, read and update test
-func TestSocialProvider_Basic(t *testing.T) {
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
-		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
-		CheckDestroy:             checksocialProviderDestroyed,
-		Steps: []resource.TestStep{
-			{
-				Config: socialProviderConfig(spName, spProviderName, spClientID, spClientSecret),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr(resourceSocialPorvider, "name", spName),
-					resource.TestCheckResourceAttr(resourceSocialPorvider, "provider_name", spProviderName),
-					resource.TestCheckResourceAttr(resourceSocialPorvider, "client_id", spClientID),
-					resource.TestCheckResourceAttr(resourceSocialPorvider, "client_secret", spClientSecret),
-					resource.TestCheckResourceAttr(resourceSocialPorvider, "claims.required_claims.user_info.0", "name"),
+// func TestSocialProvider_Basic(t *testing.T) {
+// 	resource.Test(t, resource.TestCase{
+// 		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+// 		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+// 		CheckDestroy:             checksocialProviderDestroyed,
+// 		Steps: []resource.TestStep{
+// 			{
+// 				Config: socialProviderConfig(spName, spProviderName, spClientID, spClientSecret),
+// 				Check: resource.ComposeAggregateTestCheckFunc(
+// 					resource.TestCheckResourceAttr(resourceSocialPorvider, "name", spName),
+// 					resource.TestCheckResourceAttr(resourceSocialPorvider, "provider_name", spProviderName),
+// 					resource.TestCheckResourceAttr(resourceSocialPorvider, "client_id", spClientID),
+// 					resource.TestCheckResourceAttr(resourceSocialPorvider, "client_secret", spClientSecret),
+// 					resource.TestCheckResourceAttr(resourceSocialPorvider, "claims.required_claims.user_info.0", "name"),
 
-					// default value check
-					resource.TestCheckResourceAttr(resourceSocialPorvider, "enabled", strconv.FormatBool(false)),
-					resource.TestCheckResourceAttr(resourceSocialPorvider, "enabled_for_admin_portal", strconv.FormatBool(false)),
+// 					// default value check
+// 					resource.TestCheckResourceAttr(resourceSocialPorvider, "enabled", strconv.FormatBool(false)),
+// 					resource.TestCheckResourceAttr(resourceSocialPorvider, "enabled_for_admin_portal", strconv.FormatBool(false)),
 
-					resource.TestCheckResourceAttrSet(resourceSocialPorvider, "id"),
-				),
-			},
-			{
-				ResourceName:      resourceSocialPorvider,
-				ImportState:       true,
-				ImportStateVerify: true,
-				ImportStateIdFunc: func(s *terraform.State) (string, error) {
-					rs, ok := s.RootModule().Resources[resourceSocialPorvider]
-					if !ok {
-						return "", fmt.Errorf("Not found: %s", resourceSocialPorvider)
-					}
-					return rs.Primary.Attributes["provider_name"] + ":" + rs.Primary.ID, nil
-				},
-			},
-			{
-				Config: socialProviderConfig(spName, spProviderName, spClientID, spClientSecret),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(resourceSocialPorvider, "id"),
-				),
-			},
-			// immutable name and provider_name validation
-			{
-				Config:      socialProviderConfig(spName, "facebook", spClientID, spClientSecret),
-				ExpectError: regexp.MustCompile(`Attribute 'provider_name' can't be modified`),
-			},
-			{
-				Config:      socialProviderConfig(acctest.RandString(5), spProviderName, spClientID, spClientSecret),
-				ExpectError: regexp.MustCompile(`Attribute 'name' can't be modified`),
-			},
-		},
-	})
-}
+// 					resource.TestCheckResourceAttrSet(resourceSocialPorvider, "id"),
+// 				),
+// 			},
+// 			{
+// 				ResourceName:      resourceSocialPorvider,
+// 				ImportState:       true,
+// 				ImportStateVerify: true,
+// 				ImportStateIdFunc: func(s *terraform.State) (string, error) {
+// 					rs, ok := s.RootModule().Resources[resourceSocialPorvider]
+// 					if !ok {
+// 						return "", fmt.Errorf("Not found: %s", resourceSocialPorvider)
+// 					}
+// 					return rs.Primary.Attributes["provider_name"] + ":" + rs.Primary.ID, nil
+// 				},
+// 			},
+// 			{
+// 				Config: socialProviderConfig(spName, spProviderName, spClientID, spClientSecret),
+// 				Check: resource.ComposeAggregateTestCheckFunc(
+// 					resource.TestCheckResourceAttrSet(resourceSocialPorvider, "id"),
+// 				),
+// 			},
+// 			// immutable name and provider_name validation
+// 			{
+// 				Config:      socialProviderConfig(spName, "facebook", spClientID, spClientSecret),
+// 				ExpectError: regexp.MustCompile(`Attribute 'provider_name' can't be modified`),
+// 			},
+// 			{
+// 				Config:      socialProviderConfig(acctest.RandString(5), spProviderName, spClientID, spClientSecret),
+// 				ExpectError: regexp.MustCompile(`Attribute 'name' can't be modified`),
+// 			},
+// 		},
+// 	})
+// }
 
 func socialProviderConfig(name, providerName, clientID, clientSecret string) string {
 	return fmt.Sprintf(`
 		provider "cidaas" {
-			base_url = "https://kube-nightlybuild-dev.cidaas.de"
+			base_url = "https://automation-test.dev.cidaas.eu"
 		}
 		resource "cidaas_social_provider" "example" {
 			name                     = "%s"
@@ -117,25 +113,25 @@ func socialProviderConfig(name, providerName, clientID, clientSecret string) str
 	`, name, providerName, clientID, clientSecret)
 }
 
-func checksocialProviderDestroyed(s *terraform.State) error {
-	rs, ok := s.RootModule().Resources[resourceSocialPorvider]
-	if !ok {
-		return fmt.Errorf("resource %s not fround", resourceSocialPorvider)
-	}
+// func checksocialProviderDestroyed(s *terraform.State) error {
+// 	rs, ok := s.RootModule().Resources[resourceSocialPorvider]
+// 	if !ok {
+// 		return fmt.Errorf("resource %s not fround", resourceSocialPorvider)
+// 	}
 
-	sp := cidaas.SocialProvider{
-		ClientConfig: cidaas.ClientConfig{
-			BaseURL:     os.Getenv("BASE_URL"),
-			AccessToken: acctest.TestToken,
-		},
-	}
-	res, _ := sp.Get(rs.Primary.Attributes["provider_name"], rs.Primary.Attributes["id"])
-	if res != nil {
-		// when resource exists in remote
-		return fmt.Errorf("resource stil exists %+v", res)
-	}
-	return nil
-}
+// 	sp := cidaas.SocialProvider{
+// 		ClientConfig: cidaas.ClientConfig{
+// 			BaseURL:     os.Getenv("BASE_URL"),
+// 			AccessToken: acctest.TestToken,
+// 		},
+// 	}
+// 	res, _ := sp.Get(rs.Primary.Attributes["provider_name"], rs.Primary.Attributes["id"])
+// 	if res != nil {
+// 		// when resource exists in remote
+// 		return fmt.Errorf("resource stil exists %+v", res)
+// 	}
+// 	return nil
+// }
 
 // Invalid provider_name validation
 func TestSocialProvider_InvalidProviderName(t *testing.T) {
@@ -164,7 +160,7 @@ func TestSocialProvider_MissingRequired(t *testing.T) {
 				{
 					Config: `
 						provider "cidaas" {
-							base_url = "https://kube-nightlybuild-dev.cidaas.de"
+							base_url = "https://automation-test.dev.cidaas.eu"
 						}
 						resource "cidaas_social_provider" "example" {}
 					`,